Move ska_sort to lib/ivis_opengl/3rdparty
